http server: use Flask RESTful
This change drops the ad-hoc http server and replaces it with new one
build on top of the Flask-RESTful framework. The processes API was moved
from /plugins/processes to /processes, since processes are indipendent
resource rather than subresource of the processes resource. No change
in CLI was required by this change in API.

This patch also introduces the ability of listing all processes
(GET /processes). The http server now runs in separate process which is
a step towards a state where the original process runs a watchdog
keeping an eye on the plugin and and api server subprocesses.